Add "models" for Solar System Objects:
query_close_approachers(): Queries JPL SBDB for a list of
close-approaching asteroids which meet various criteria. Results of this
can be passed to the next routine..
get_horizons_ephem(): Routine to query JPL Horizons for an ephmeris for
Rubin for an adjustable amount of days around close approach.
transform_Vmag(): Perform V mag to (PanSTARRS but close to Rubin) grizYw filter transformations for a variety of asteroid taxonomic classes including a mean S+C set which represents ~65% of NEO population.